requirements.txt
regex~=2024.9.11
python-docx
